Textfield not required

+++ api/views.py
@@ -4,7 +4,8 @@ from api.models import Main,OtherField
 from pymongo import MongoClient
 from Ostadkar_final.settings import HOST_NAME,DATABASE_NAME
 import json
-# Create your views here.
+
+
 def main(request):
     mform = MainForm()
     oform = OtherForm()
@@ -18,25 +19,30 @@ def main(request):
         mobj.priority = mform['priority'].value()
         mobj.description = mform['description'].value()
         oobj.other = oform['other'].value()
-        oobj.save()
-        mobj.other_fk_id = oobj.id
+        if oobj.other:
+            oobj.save()
+            mobj.other_fk_id = oobj.id
+
+
         print(mobj.other_fk_id)
         if mform.is_valid() and oform.is_valid():
+            # Connection
             myclient = MongoClient(HOST_NAME)
             mydb = myclient[DATABASE_NAME]
             mycol = mydb["api_otherfield"]
-            jsonfile = json.loads(oobj.other)
-            oobj.save()
-            jsonfile['id'] = oobj.id
-            mobj.other_fk_id = oobj.id
-            mycol.delete_one({"id":oobj.id})
-            mycol.insert_one(jsonfile)
 
-            # mobj.other_fk_id = x.inserted_id
-
+            if oobj.other:
+                try :
+                    jsonfile = json.loads(oobj.other)
+                    jsonfile['id'] = oobj.id
+                    mobj.other_fk_id = oobj.id
+                    mycol.delete_one({"id":oobj.id})
+                    mycol.insert_one(jsonfile)
+                except:
+                    print("Could not load json")
 
             mobj.save()
-
-
     return render(request,'api/main.html', context={'form1':mform,'form2':oform})
+
+
